\name{beta_0.75_0.65}
\docType{data}
\alias{beta_0.75_0.65}

\title{Synthetic dataset from a beta distribution}
\description{
This is a synthetic generated dataset sampling a beta distribution with parameters \code{shape1 = 0.75} and \code{shape2 = 0.65} 
}
\usage{beta_0.75_0.65}
\format{A vector containing 10000 observations.}

\keyword{datasets}