\name{beta_5_10}
\docType{data}
\alias{beta_5_10}

\title{Synthetic dataset from a beta distribution}
\description{
This is a synthetic generated dataset sampling a beta distribution with parameters \code{shape1 = 5} and \code{shape2 = 10} 
}
\usage{beta_5_10}
\format{A vector containing 10000 observations.}

\keyword{datasets}