\name{eruption}
\docType{data}
\alias{eruption}

\title{Eruption lengths of Old Faithful geyser}
\description{
The dataset comprises lengths (in minutes) of eruptions of Old Faithful geyser in Yellowstone National Park, USA. The data are within the interval [1.67,4.93].
}
\usage{eruption.r}
\format{A vector containing 107 observations.}
\source{The data were obtained from Silverman (1996) Table 2.2}

\references{
Silverman, B. (1986). \emph{Density Estimation for Statistics and Data Analysis}. 
Chapman & Hall

Weisberg, S. (1980). \emph{Applied linear regression}.
John Wiley & Sons, Canada
}

\keyword{datasets}