\name{getb}
\alias{getb}
\alias{getb-methods}
\alias{getb,KernelDensity-method}
\alias{getb,Chen99Kernel-method}
\alias{getb,MicroBetaChen99Kernel-method}
\alias{getb,MacroBetaChen99Kernel-method}
\alias{getb,NoBoundaryKernel-method}
\alias{getb,Muller91BoundaryKernel-method}
\alias{getb,Muller94BoundaryKernel-method}
\alias{getb,NormalizedBoundaryKernel-method}
\alias{getb,HirukawaJLNKernel-method}
\alias{getb,HirukawaTSKernel-method}
\alias{getb,MacroBetaHirukawaJLNKernel-method}
\alias{getb,MacroBetaHirukawaTSKernel-method}

\title{Accesor method for \code{b} slot}
\description{
 This method obtains the values stored in the \code{b} slot of a bounded density object. This slot contains the bandwidth parameter for the kernel estimator.
}

\section{Methods}{
\describe{

\item{\code{getb(.Object)}}{
%%  ~~describe this method here~~
}
}}

\arguments{
\item{\code{.Object}}{A kernel density estimator.  See all the accepted classes here by running the command \code{getSubclasses("KernelDensity")}.}
}

\keyword{methods}

