\name{getc}
\alias{getc}
\alias{getc-methods}
\alias{getc,HirukawaTSKernel-method}
%\alias{getc,MacroBetaHirukawaTSKernel-method}

\title{Accesor method for \code{c} slot}
\description{
 This method obtains the values stored in the \code{c} slot of a HirukawaTSKernel object. This parameter is used in the kernel estimation as a smoothing parameter.
}

\section{Methods}{
\describe{

\item{\code{getc(.Object)}}{
%%  ~~describe this method here~~
}
}}

\arguments{
\item{\code{.Object}}{A \code{\linkS4class{HirukawaTSKernel}} or a \code{\linkS4class{MacroBetaHirukawaTSKernel}} object.}
}

\keyword{methods}

