\name{getdataPointsCache}
\alias{getdataPointsCache}
\alias{getdataPointsCache-methods}
\alias{getdataPointsCache,BoundedDensity-method}
\alias{getdataPointsCache,Chen99Kernel-method}
\alias{getdataPointsCache,MicroBetaChen99Kernel-method}
\alias{getdataPointsCache,MacroBetaChen99Kernel-method}
\alias{getdataPointsCache,NoBoundaryKernel-method}
\alias{getdataPointsCache,Muller91BoundaryKernel-method}
\alias{getdataPointsCache,Muller94BoundaryKernel-method}
\alias{getdataPointsCache,NormalizedBoundaryKernel-method}
\alias{getdataPointsCache,Vitale-method}
\alias{getdataPointsCache,BrVitale-method}
\alias{getdataPointsCache,KakizawaB1-method}
\alias{getdataPointsCache,KakizawaB2-method}
\alias{getdataPointsCache,KakizawaB3-method}
\alias{getdataPointsCache,HirukawaJLNKernel-method}
\alias{getdataPointsCache,HirukawaTSKernel-method}
\alias{getdataPointsCache,MacroBetaHirukawaJLNKernel-method}
\alias{getdataPointsCache,MacroBetaHirukawaTSKernel-method}

\title{Accesor method for \code{DataPointsCache} slot}
\description{
 This method obtains the values stored in the \code{dataPointsCache} slot of a bounded density object.
}

\section{Methods}{
\describe{

\item{\code{getdataPointsCache(.Object)}}{
%%  ~~describe this method here~~
}
}}

\arguments{
\item{\code{.Object}}{A bounded density estimator.  See all the accepted classes here by running the command \code{getSubclasses("BoundedDensity")}.}
}

\keyword{methods}

