\name{getdensityCache}
\alias{getdensityCache}
\alias{getdensityCache-methods}
\alias{getdensityCache,BoundedDensity-method}
\alias{getdensityCache,Chen99Kernel-method}
\alias{getdensityCache,MicroBetaChen99Kernel-method}
\alias{getdensityCache,MacroBetaChen99Kernel-method}
\alias{getdensityCache,NoBoundaryKernel-method}
\alias{getdensityCache,Muller91BoundaryKernel-method}
\alias{getdensityCache,Muller94BoundaryKernel-method}
\alias{getdensityCache,NormalizedBoundaryKernel-method}
\alias{getdensityCache,Vitale-method}
\alias{getdensityCache,BrVitale-method}
\alias{getdensityCache,KakizawaB1-method}
\alias{getdensityCache,KakizawaB2-method}
\alias{getdensityCache,KakizawaB3-method}
\alias{getdensityCache,HirukawaJLNKernel-method}
\alias{getdensityCache,HirukawaTSKernel-method}
\alias{getdensityCache,MacroBetaHirukawaJLNKernel-method}
\alias{getdensityCache,MacroBetaHirukawaTSKernel-method}

\title{Accesor method for \code{densityCache} slot}
\description{
This method obtains the values stored in the \code{DensityCache} slot of a bounded density object
}
\section{Methods}{
\describe{
\item{\code{getdensityCache(.Object)}}{
%%  ~~describe this method here~~
}
}}

\arguments{
\item{\code{.Object}}{A bounded density estimator.  See all the accepted classes here by running the command \code{getSubclasses("BoundedDensity")}.}
}

\keyword{methods}
