\name{getdensityEstimator}
\alias{getdensityEstimator}
\alias{getdensityEstimator-methods}
\alias{getdensityEstimator,KakizawaB1-method}
\alias{getdensityEstimator,KakizawaB2-method}
\alias{getdensityEstimator,KakizawaB3-method}


\title{Accesor method for \code{gamma} slot}
\description{
 This method obtains the class name of the object stored in the \code{densityEstimator} slot of a KakizawaB1, KakizawaB2 or KakizawaB3 object.
}

\section{Methods}{
\describe{

\item{\code{getdensityEstimator(.Object)}}{
%%  ~~describe this method here~~
}
}}

\arguments{
\item{\code{.Object}}{A \code{\linkS4class{KakizawaB1}}, \code{\linkS4class{KakizawaB2}} or \code{\linkS4class{KakizawaB3}} object.}
}

\keyword{methods}

