\name{getdistributionCache}
\alias{getdistributionCache}
\alias{getdistributionCache-methods}
\alias{getdistributionCache,BoundedDensity-method}
\alias{getdistributionCache,Chen99Kernel-method}
\alias{getdistributionCache,MicroBetaChen99Kernel-method}
\alias{getdistributionCache,MacroBetaChen99Kernel-method}
\alias{getdistributionCache,NoBoundaryKernel-method}
\alias{getdistributionCache,Muller91BoundaryKernel-method}
\alias{getdistributionCache,Muller94BoundaryKernel-method}
\alias{getdistributionCache,NormalizedBoundaryKernel-method}
\alias{getdistributionCache,Vitale-method}
\alias{getdistributionCache,BrVitale-method}
\alias{getdistributionCache,KakizawaB1-method}
\alias{getdistributionCache,KakizawaB2-method}
\alias{getdistributionCache,KakizawaB3-method}
\alias{getdistributionCache,HirukawaJLNKernel-method}
\alias{getdistributionCache,HirukawaTSKernel-method}
\alias{getdistributionCache,MacroBetaHirukawaJLNKernel-method}
\alias{getdistributionCache,MacroBetaHirukawaTSKernel-method}

\title{Accesor method for \code{distributionCache} slot}
\description{
 This method obtains the values stored in the \code{DistributionCache} slot of a bounded density object.
}

\section{Methods}{
\describe{

\item{\code{getdistributionCache(.Object)}}{
%%  ~~describe this method here~~
}
}}

\arguments{
\item{\code{.Object}}{A bounded density estimator.  See all the accepted classes here by running the command \code{getSubclasses("BoundedDensity")}.}
}

\keyword{methods}

