\name{getgamma}
\alias{getgamma}
\alias{getgamma-methods}
\alias{getgamma,KakizawaB1-method}
\title{Accesor method for \code{gamma} slot}
\description{
 This method obtains the values stored in the \code{gamma} slot of a KakizawaB1 object. This slot contains a parameter used in the B1 approximation using Bernstein polynomials.
}

\section{Methods}{
\describe{

\item{\code{getgamma(.Object)}}{
%%  ~~describe this method here~~
}
}}

\arguments{
\item{\code{.Object}}{A \code{\linkS4class{KakizawaB1}} object.}
}

\keyword{methods}

