\name{getM}
\alias{getM}
\alias{getM-methods}
\alias{getM,BernsteinPolynomials-method}
\alias{getM,BrVitale-method}
\title{Accesor method for \code{M} slot}
\description{
 This method obtains the values stored in the \code{M} slot of a BrVitale object. This slot contains parameter for bias reduction.
}

\section{Methods}{
\describe{

\item{\code{getM(.Object)}}{
%%  ~~describe this method here~~
}
}}

\arguments{
\item{\code{.Object}}{A \code{\linkS4class{BrVitale}} Object.}
}

\keyword{methods}

