\name{getm}
\alias{getm}
\alias{getm-methods}
\alias{getm,BernsteinPolynomials-method}
\alias{getm,Vitale-method}
\title{Accesor method for \code{m} slot}
\description{
 This method obtains the values stored in the \code{m} slot of a BernsteinPolynomials object. This slot contains the order of the polynomial expansion.
}

\section{Methods}{
\describe{

\item{\code{getm(.Object)}}{
%%  ~~describe this method here~~
}
}}

\arguments{
\item{\code{.Object}}{A boundary kernel density estimator. See all the accepted classes here with  \code{getSubclasses("BernsteinPolynomials")}.}
}

\keyword{methods}

