\name{getmodified}
\alias{getmodified}
\alias{getmodified-methods}
\alias{getmodified,Chen99Kernel-method}
\alias{getmodified,MicroBetaChen99Kernel-method}
\alias{getmodified,MacroBetaChen99Kernel-method}
\alias{getmodified,HirukawaJLNKernel-method}
\alias{getmodified,HirukawaTSKernel-method}
\alias{getmodified,MacroBetaHirukawaJLNKernel-method}

\title{Accesor method for \code{modified} slot}
\description{
 This method obtains the values stored in the \code{modified} slot of a Kernel density object. The value of this slot is \code{TRUE} if a modified version of the kernel estimator is used and \code{FALSE} otherwise.
}

\section{Methods}{
\describe{

\item{\code{getgetmodified(.Object)}}{
%%  ~~describe this method here~~
}
}}

\arguments{
\item{\code{.Object}}{A kernel density estimator.  See all the accepted classes here by running the command \code{getSubclasses("KernelDensity")}.}
}

\keyword{methods}

