\name{lines}
\alias{lines}
\alias{lines-methods}
\alias{lines,BoundedDensity-method}
\alias{lines,KernelDensity-method}
\alias{lines,Chen99Kernel-method}
\alias{lines,MicroBetaChen99Kernel-method}
\alias{lines,MacroBetaChen99Kernel-method}
\alias{lines,BoundaryKernel-method}
\alias{lines,NoBoundaryKernel-method}
\alias{lines,Muller91BoundaryKernel-method}
\alias{lines,JonesCorrectionMuller91BoundaryKernel-method}
\alias{lines,Muller94BoundaryKernel-method}
\alias{lines,JonesCorrectionMuller94BoundaryKernel-method}
\alias{lines,NormalizedBoundaryKernel-method}
\alias{lines,BernsteinPolynomials-method}
\alias{lines,Vitale-method}
\alias{lines,BrVitale-method}
\alias{lines,KakizawaB1-method}
\alias{lines,KakizawaB2-method}
\alias{lines,KakizawaB3-method}
\alias{lines,HirukawaJLNKernel-method}
\alias{lines,HirukawaTSKernel-method}
\alias{lines,MacroBetaHirukawaJLNKernel-method}
\alias{lines,MacroBetaHirukawaTSKernel-method}

\title{Add a Bounded Density pdf to a Plot}
\description{
 Function to draw a bounded density probability density functions in the current plot.
}
\section{Methods}{
\describe{

\item{\code{lines(x,...)}}{
%%  ~~describe this method here~~
}
}}
\arguments{
\item{\code{x}}{A bounded density estimator.See all the accepted classes here by running the command \code{getSubclasses("BoundedDensity")}.}
\item{\code{...}}{Arguments to be passed to methods, such as graphical parameters (see \code{\link{par}}).}
}
\keyword{methods}
