\name{quantile}
\alias{quantile}
\alias{quantile-methods}
\alias{quantile,BoundedDensity-method}
\alias{quantile,Chen99Kernel-method}
\alias{quantile,MicroBetaChen99Kernel-method}
\alias{quantile,MacroBetaChen99Kernel-method}
\alias{quantile,BoundaryKernel-method}
\alias{quantile,NoBoundaryKernel-method}
\alias{quantile,Muller91BoundaryKernel-method}
\alias{quantile,JonesCorrectionMuller91BoundaryKernel-method}
\alias{quantile,Muller94BoundaryKernel-method}
\alias{quantile,JonesCorrectionMuller94BoundaryKernel-method}
\alias{quantile,NormalizedBoundaryKernel-method}
\alias{quantile,Vitale-method}
\alias{quantile,BrVitale-method}
\alias{quantile,KakizawaB1-method}
\alias{quantile,KakizawaB2-method}
\alias{quantile,KakizawaB3-method}
\alias{quantile,HirukawaJLNKernel-method}
\alias{quantile,HirukawaTSKernel-method}
\alias{quantile,MacroBetaHirukawaJLNKernel-method}
\alias{quantile,MacroBetaHirukawaTSKernel-method}


\title{Quantile}
\description{
 Quantile function for the given bounded density object.
}
\section{Methods}{
\describe{

\item{\code{quantile(x,p)}}{
%%  ~~describe this method here~~
}
}}
\arguments{
\item{\code{x}}{A bounded density estimator. See all the accepted classes here by running the command \code{getSubclasses("BoundedDensity")}.This parameter is named \code{x} instead of \code{.Object} to agree with other already defined density methods.}
\item{\code{p}}{Vector of probabilities}
}
\keyword{methods}
%\keyword{ ~~ other possible keyword(s) ~~ }
