\name{rsample}
\alias{rsample}
\alias{rsample-methods}
\alias{rsample,BoundedDensity-method}
\alias{rsample,Chen99Kernel-method}
\alias{rsample,MicroBetaChen99Kernel-method}
\alias{rsample,MacroBetaChen99Kernel-method}
\alias{rsample,BoundaryKernel-method}
\alias{rsample,NoBoundaryKernel-method}
\alias{rsample,Muller91BoundaryKernel-method}
\alias{rsample,JonesCorrectionMuller91BoundaryKernel-method}
\alias{rsample,Muller94BoundaryKernel-method}
\alias{rsample,JonesCorrectionMuller94BoundaryKernel-method}
\alias{rsample,NormalizedBoundaryKernel-method}
\alias{rsample,Vitale-method}
\alias{rsample,BrVitale-method}
\alias{rsample,KakizawaB1-method}
\alias{rsample,KakizawaB2-method}
\alias{rsample,KakizawaB3-method}
\alias{rsample,HirukawaJLNKernel-method}
\alias{rsample,HirukawaTSKernel-method}
\alias{rsample,MacroBetaHirukawaJLNKernel-method}
\alias{rsample,MacroBetaHirukawaTSKernel-method}


\title{Random sample}
\description{
 Random generator function for the given bounded density object.
}
\section{Methods}{
\describe{

\item{\code{rsample(.Object,n)}}{
%%  ~~describe this method here~~
}
}}
\arguments{
\item{\code{.Object}}{A bounded density estimator. See all the accepted classes here by running the command \code{getSubclasses("BoundedDensity")}.}
\item{\code{n}}{number of random observations to be generated}
}
\keyword{methods}
%\keyword{ ~~ other possible keyword(s) ~~ }
