\name{suicide.r}
\docType{data}
\alias{suicide.r}

\title{Scaled data from suicide risk data}
\description{
The dataset comprises lengths (in days) of psychiatric treatment spells for patients used as controls in a study of suicide risks. The data have been scaled to the interval [0,1] by dividing each data sample by the maximum value.
}

\usage{suicide.r}
\format{A vector containing 86 observations.}
\source{The data were obtained from Silverman (1996) Table 2.1}

\references{
Silverman, B. (1986). \emph{Density Estimation for Statistics and Data Analysis}. 
Chapman & Hall

Copas, J. B. and Fryer, M. J. (1980). Density estimation and suicide risks in psychiatric treatment.
\emph{Journal of the Royal Statistical Society. Series A}, 143(2), 167-176
}

\keyword{datasets}