\name{tgaussian}
\docType{data}
\alias{tgaussian}

\title{Synthetic dataset from a truncated Gaussian distribution}
\description{
This is a synthetic generated dataset sampling a truncated Gaussian distribution on the interval [0,1] with \code{mean=0} and \code{sd=0.25}
}
\usage{tgaussian}
\format{A vector containing 10000 observations.}

\keyword{datasets}